==> analyses/asd_analyses/single-cell_rna/2_analyses/0.Common.ipynb <==
library(dplyr)
library(RColorBrewer)

sapply(phases.colors, function(x){a <- col2rgb(x)
                                  rgb(a[1],a[2],a[3], maxColorValue = 255)})

dir.create('../results/')
dir.create('../plots/')

### **FUNCTIONS**

# Sets width and height of figures (for jupyter notebooks)
fig <- function(w, h){
    options(repr.plot.width=w, repr.plot.height=h)
}

### **ANTEROVENTRAL GENES**
Genes with anteroventral bias coming from Fig XX from the Monkey paper [href]

monkey.markers <- list(
    'RPC' = c(
        'FGF17', 'FGF18', 'FGF3', 'SP8', 'FGF8', 'DLK1', 'SPRY2', 'ZIC4', 'ZIC3', 'ETV1', 'ZIC1'
    ),
    'AV' = c(
        'NKX6-2', 'SHH', 'SOX3', 'NKX2-1', 'FOXG1', 'DLX2', 'OLIG1', 'OLIG2', 'DLK1', 'SOX6', 'ASCL1' #MANUALLY ADDED
    ),
    'GE' = c(
        'MEIS2', 'POU3F3', 'FOXG1', 'GSX2', 'PBX3', 'OLIG2', 'POU3F2', 'ZEB2', 'ZEB1', 'FGFR2', 'MAP2'
    ),
    'forebrain' = c('FOXG1'),
    'ventral'=c('GSX2', 'NKX2-1'), 
    'anteroventral'=c('SIX3', 'SP8'),
    'ventrolateral'=c('LMO3', 'NRP2'), 
    'anti-hem'=c('FGF7', 'SFRP2', 'SFRP1'), 
    'dorsolateral'=c('EMX1', 'LHX2', 'NEUROG2'), 
    'anteromedial'=c('NEUROG2'), 
    'caudomedial'=c('LEF1', 'NR2F1', 'EMX2', 'FGFR3'),
    'hem/medial'=c('LMX1A', 'WNT5A', 'WNT8B', 'RSPO3', 'ZIC2', 'OTX2', 'ZNF521'), 
    'choroid pl.'=c('TTR')
)

monkey.genes <- unlist(monkey.markers) %>% unique()

list('Monkey named genes'=c(
    'SOX2', 'NES', 'FGF8', 'FGF17', 'FGF18', 'SP8', 'FOXG1',
    'NKX2-1', 'SHH', 'NKX6-2', 'LMO1', 'ZIC1', 'ZIC3', 'ZIC4',
    'LMX1A', 'WNT', 'RSPO3', 'WNT8B', 'BAMBI', 'RSPO3', 'TTR', 
    'ARX', 'FGFR3', 'LHX9', 'TCF7L2', 'SFRP2','FGF3', 'FGF17', 
    'FGF18', 'SPRY2', 'ZNF219', 'NFIX', 'ARX', 'LHX9', 'RGMA', 
    'FSTL1', 'LEFT1', 'LHX8', 'ONECUT1', 'ONECUT2', 'TP73', 'FGF17', 'FGFR1', 'FGFR3', 
    'WNT5A', 'FZD5', 'WNT5A', 'FZD5', 'WNT5A', 'GALP', 'GALR2', 'GAL', 
    'RSPO3', 'SP8', 'KI67', 'SOX2', 'RSPO3', 'EDU', 'SOX2', 'PMP22', 
    'CYP6A1', 'HMGA2', 'JARID2', 'FOXJ1', 'VGFA', 'ZNF219', 'ZIC5', 'SOX21', 
    'WNT7B', 'WNT8B', 'ZYP26A1', 'RBP1', 'HOPX', 'PTN', 'FABP7',
    'PMP22', 'NR2F1', 'NR2F2', 'HES5', 'FGFR3', 'PENK', 'RBP1', 
    'MEF2C', 'NPY', 'NTRK2', 'RBP1', 'DCT',
    'SP8', 'LHX9', 'FEZF1', 'NR2F1', 'MEIS2', 'PBX1',
    'EOMES', 'SOX5', 'CUX2', 'NR4A2', 'GRID2', 'SYT6', 'OPRK1', 'PCDH10', 'PCDH17', 'CYP26A1',
    'CBLN2', 'MEIS2', 'BLC6', 'CBLN2', 'RBP1', 'NKX2', 
    'NKX1', 'PMP22', 'HOPX', 'CYP26A1', 'RBP1', 'NR2F1', 'RSPO3', 'NR2F2', 'CYP26B1', 
    'CYP26A1', 'BRINP1', 'CBLN2', 'LFRN2', 'CAMKV', 'LHX6',
    'NR2F2', 'SP8', 'LHX6', 'NR2F2', 'EFGR', 'GFAP',
    'MFGE8', 'EGFR', 'HMGN3', 'HMGB2', 'PAX6', 'HES1', 'SOX2', 
    'SOX3', 'SOX6', 'OLIG1', 'OLIG2', 'ASCL1', 'SOX10', 'NKX2-2', 'SLC35F1', 'STAT3', 'JAG1', 'JAG2',
    'DLL4', 'MDK', 'NOTCH1', 'NOTCH2', 'NOTCH3', 'DC2', 'ITGB1', 'CDON', 'SATB1', 'THSB1', 'HES1', 'CBLN2')
)%>% lapply(unique) %>% lapply(sort) -> monkey_paper_list


nicola_marker_list <- list(
    'Early AV Prog / vNe' = c('FGF8', 'FOXG1', 'FZD5', 'SIX3', 'DUSP6', 'TET1', 'PBX1', 'HES5', 'HES4', 'LAMP5', 'CAMK2N1'),
    'Frontal' = c('RBP1', 'PTN'),
    'Neurons' = c('ONECUT2'),
    'Dorsal' = c('FEZF2', 'DMRTA2', 'WLS', 'NR2F1'),
    'Posterior' = c('PEG3', 'ANXA2')
)

shown_4d_list <- list(
    'FR' = c('SP8', 'WNT7B', 'ZIC5', 'WNT8B', 'ETV4', 'RBP1', 'ZIC4','EFNA5', 'MEIS3', 'ZNF219', 'HES5', 'HMGA1', 'ZIC2', 'NR2F6', 'CYP26A1', 'ZBF775', 'FOXG1', 'ZIC1', 'ZIC3', 'MAP2', 'ETV5', 'PTN', 'CXXC5', 'SOX21', 'HOPX', 'DKK3', 'ETV1', 'FABP7'),
    'MS' = c('HEY1', 'SOX3', 'EMX2', 'LAMTOR3', 'CREB5', 'RAB7A', 'DKK3'),
    'OC' = c('RSPO3', 'FGF13', 'SP5', 'IGF2', 'ZBTB16', 'NR2F1', 'FGFR3', 'NR2F2', 'HES5', 'FEZF1', 'HEY1', 'ZNF219'),
    'Tem' = c('PEG3', 'NR2F1', 'FGFR3', 'NR2F2', 'RSPO3')
)


shown_2b_list <- list(
    'RPC' = c('FGF17',
              'FGF18',
              'FGF3',
              'SP8',
              'FGF8',
              'DLK1',
              'SPRY2',
              'ZIC4',
              'ZIC3',
              'ETV1',
              'ZIC1'),
    'AV'= c('NKX6-2',
            'SHH',
            'SOX3',
            'NKX2-1',
            'FOXG1',
            'DLX2',
            'OLIG1',
            'OLIG2',
            'DLK1',
            'SOX6',
            'ASCL1'),
    'GE'=c('MEIS',
           'POU3F3',
           'FOXG1',
           'GSX2',
           'PBX3',
           'OLIG2',
           'POU3F2',
           'ZEB2',
           'ZEB1',
           'FGFR2',
           'MAP2'),
    'Hem'= c('RSPO2',
             'WNT2B',
             'WNT8B',
             'CALB1',
             'RSPO3',
             'LHX9',
             'SP5',
             'DMRTA2',
             'EMX2',
             'FGF13',
             'ARX',
             'BAMBI',
             'FGFR3'),
    'Hem/CPe' = c('CA2',
                  'TTR',
                  'NPY1R',
                  'FOXJ1',
                  'BMP4'),
    'Anti-Hem' = c('SFRP2',
                   'PAX6',
                   'NR2F2',
                   'NR2F1'),
    'ZLI' = c('WNT4',
              'IRX3',
              'WNT3',
              'OTX1',
              'TCF7L2',
              'GLI2')
)
    


list(
    RGC=c('BLBP', 'NES', 'PAX6', 'POU3F2', 'BRN2', 'SOX1', 'SOX2'),
    ImNeu=c('DCX','TUBB3'), 
    Neu=c('NEU1','NEU2','NEU3','NEU4'), 
    vRGC=c('GAD1'), 
    AMPArec=c('GRIA1', 'GRIA2', 'GRIA3', 'GRIA4'), 
    ASD_cellfate_prolif=c('DLX6-AS1', 'FOXG1', 'EOMES', 'POU3F3', 'BRN1', 'SOX3', 'SOX5', 'GSX2', 'ETV1', 'DLX1', 'DLX6', 'E2F2', 'SYNE2'), 
    ASD_neuronal_diff=c('NCAM1', 'NRCAM', 'L1CAM', 'NFASC', 'PLXNC1', 'PLXNB3', 'DCC', 'UNC5A', 'PAK3', 'PAK7', 'DCK5R1', 'P35', 'DCX', 'CDK5R1', 'DPYSL5', 'DCX', 'DPF1', 'APC2'), 
    ASD_synap_ion_chs=c('NRXN1', 'NRXN2', 'SLITRK1', 'CAMK2B', 'CAMK1D', 'NRSN1', 'SYT13', 'GRIN1', 'SCN2A', 'TSPAN7'), 
    Glut=c('SLC32A1', 'GAD1', 'GAD67'), 
    ASD_other=c('ASCL1', 'MASH1', 'NKX2-1', 'NKX2.1', 'GSX2', 'ASCL1', 'DLX1', 'DLX2', 'DLX5', 'DLX6', 'DLX6-AS1'), 
    FOXG1_rel=c('FOXG1', 'DLX6-AS1', 'TMEM132C', 'C14orf23', 'KLHDC8A')
) %>% lapply(unique) %>% lapply(sort) -> flora_paper_list

openxlsx::getSheetNames('../raw/monkey_markers/table S2.subtype.markers.xlsx')
top100.subtype <- as.list(as.data.frame(t(openxlsx::read.xlsx('../raw/monkey_markers/table S2.subtype.markers.xlsx', startRow = 3, rowNames = T, colNames = F))))

names(top100.subtype) <- gsub(names(top100.subtype), pattern='AntVen', replacement='AV', fixed=T)

openxlsx::getSheetNames('../raw/monkey_markers/table S3 organizer_markers.xlsx')
pat.de <- openxlsx::read.xlsx('../raw/monkey_markers/table S3 organizer_markers.xlsx', startRow = 2)
# pat.de <- pat.de %>% subset(Adjusted.P.value < 0.05)
lapply(split(pat.de$gene, pat.de$Subtype), length)
pat.list <- split(pat.de$gene, pat.de$Subtype)

pat.de %>% group_by(Subtype)%>% arrange(desc(Ratio.fold.changes))%>%slice_head(n=20) -> pat.de.top20
pat.top20 <- split(pat.de.top20$gene, pat.de.top20$Subtype)
lapply(pat.top20, length)

openxlsx::getSheetNames('../raw/monkey_markers/table S4 organizer markers versus NSCs.xlsx')
pat.neu.de <- openxlsx::read.xlsx('../raw/monkey_markers/table S4 organizer markers versus NSCs.xlsx', startRow = 2)
pat.neu.de <- pat.neu.de %>% mutate(Subtype=gsub(pattern='AntVen', replacement='AV', Subtype))
lapply(split(pat.neu.de$gene, pat.neu.de$Subtype), length)
pat.neu.list <- split(pat.neu.de$gene, pat.neu.de$Subtype)

pat.neu.de %>% group_by(Subtype)%>% arrange(desc(Ration.fold.changes))%>%slice_head(n=20) -> pat.neu.de.top20
pat.neu.top20 <- split(pat.neu.de.top20$gene, pat.neu.de.top20$Subtype)
lapply(pat.neu.top20, length)


top100.sub.pat <- top100.subtype[unique(pat.neu.de$Subtype)]

region.markers <- openxlsx::read.xlsx('../raw/adf3786_Suppl. Excel_seq37_v1.xlsx', startRow = 2)
colnames(region.markers)[-1] <- sapply(strsplit(colnames(region.markers)[-1], split='.', fixed=T), function(x){paste(x[-(1:3)], collapse='_')})
region.list <- apply(region.markers[,-1], 2, function(x){region.markers[x==1,1]})
region.genes <- unique(unlist(region.list))
sapply(region.list, length)

nescreg.markers <- openxlsx::read.xlsx('../raw/science.adf3786_table_s6.xlsx', startRow = 2) %>% arrange(Enriched.region, `Pseudotime.on.(min.set.to.-100)`, `Pseudotime.off.(min.set.to.100)`)

nescreg.list <- split(nescreg.markers$Gene, nescreg.markers$Enriched.region) %>% lapply(unique)
nescreg.genes <- unique(unlist(nescreg.list))

nescreg.markers.no_trg <- nescreg.markers[,-7]
nescreg.markers.no_trg <-nescreg.markers.no_trg[rowSums(nescreg.markers.no_trg[,4:6])!=0,]
nescreg.list.no_trg <- split(nescreg.markers.no_trg$Gene, nescreg.markers.no_trg$Enriched.region) %>% lapply(unique)
nescreg.genes.no_trg <- unique(unlist(nescreg.list.no_trg))


nescreg.markers.no_trg.early <- subset(nescreg.markers.no_trg, `Pseudotime.on.(min.set.to.-100)`<0)
nescreg.list.no_trg.early <- split(nescreg.markers.no_trg.early$Gene, nescreg.markers.no_trg.early$Enriched.region) %>% lapply(unique)
nescreg.genes.no_trg.early <- unique(unlist(nescreg.list.no_trg.early))


# nescreg.markers[rowSums(nescreg.markers[, 4:6])>0,]
# lapply(4:6, function(x){
#     a <- nescreg.markers[nescreg.markers[,x]==1,]
#     a <- split(a$Gene, a$Enriched.region)
#     names(a) <- paste(names(a), gsub(pattern='Enriched.in.', fixed=T, replacement='', colnames(nescreg.markers)[x]))
#     a
# }) %>% do.call(what='c') -> nescreg.ct.list

### **TRANSCRIPTION FACTORS**

tfs.df <- read.csv('~/codebases/cortical_disorders2/raw/HumanTFs/DatabaseExtract_v_1.01.csv', row.names=1)
tfs.df <- subset(tfs.df, grepl('Yes', Is.TF.))
tf.genes <- unique(tfs.df$HGNC.symbol)

tf.genes <- stringr::str_replace(tf.genes, pattern = ' ', replace='')

### **IMPRINTING**

imprinted.df <- read.csv('~/codebases/ASD_IPSC_june/imprinted_human_genes.tsv', sep='\t') %>% 
    mutate(gene=paste(Gene, Aliases, sep=', '), Aliases=NULL, Location=NULL, Gene=NULL) %>%
    rename(Imprinted.Status=Status) %>%
    tidyr::separate_rows(gene, sep = ', ') %>% arrange_all() %>% subset((gene!=' ') & !duplicated(gene))

imprinted.genes <- unique(imprinted.df$gene)

### **SEX CHROMOSOME GENES**

library(biomaRt)

valid_chr <- c(as.character(1:22), 'X', 'Y', 'MT')

mart <- useEnsembl(biomart="ensembl", dataset="hsapiens_gene_ensembl", mirror = 'asia')

chr_df <- getBM(attributes = c("hgnc_symbol", 'ensembl_gene_id', "chromosome_name"), mart = mart) %>% arrange_all()

chr_df$hgnc_symbol <- plyr::mapvalues(
    x = chr_df$ensembl_gene_id,
    from = subset(chr_df, nchar(hgnc_symbol) > 1)$ensembl_gene_id,
    to = subset(chr_df, nchar(hgnc_symbol) > 1)$hgnc_symbol) %>% make.unique()

chr_df <- subset(chr_df, (chromosome_name%in%valid_chr)&(nchar(hgnc_symbol)>1)) %>% mutate(ensembl_gene_id=NULL)
chr_df %>% arrange(factor(chromosome_name, valid_chr)) -> chr_df

sex_chr.genes <- sort(unique(subset(chr_df, chromosome_name %in% c('X','Y'))$hgnc_symbol))
sex_chr.genes.x <- sort(unique(subset(chr_df, chromosome_name == 'X')$hgnc_symbol))
sex_chr.genes.y <- sort(unique(subset(chr_df, chromosome_name == 'Y')$hgnc_symbol))

chr_df %>% group_by(hgnc_symbol) %>% summarise(chr=paste(chromosome_name, collapse=',')) -> chr_df

write.table(chr_df, file = '../data/gene_chromosomes.txt')

table(X=sex_chr.genes%in%sex_chr.genes.x, Y=sex_chr.genes%in%sex_chr.genes.y)

### **GENE TYPE**

gene_type.df <- getBM(attributes = c("hgnc_symbol", 'ensembl_gene_id', "gene_biotype"), mart = mart) %>% arrange_all()

gene_type.df$hgnc_symbol <- plyr::mapvalues(
    x = gene_type.df$ensembl_gene_id,
    from = subset(gene_type.df, nchar(hgnc_symbol) > 1)$ensembl_gene_id,
    to = subset(gene_type.df, nchar(hgnc_symbol) > 1)$hgnc_symbol) %>% make.unique()


gene_type.df <- subset(gene_type.df, (nchar(hgnc_symbol)>1)) %>% mutate(
    ensembl_gene_id=NULL,
    protein_coding=grepl(pattern='protein_coding', gene_biotype))

protein_coding.genes <- sort(unique(subset(gene_type.df, protein_coding)$hgnc_symbol))

table(grepl(protein_coding.genes, pattern='^ENSG'))

### **FACTORS**

# Cell types in the data
cell.types <- c('FGF17-like', 'Mes prog', 'RG early', 'RG late', 'Neral crest/Mes', 'Neuron', 'High-mito')
cell.types.nohighmito <- cell.types[cell.types != 'High-mito']

# Cell cycle phases
phases <- c('G1', 'S', 'G2M')

# Donors
# samples <- sort(unique(as.character(seu$cell_origin)))
samples <- c('D8_ASD_375','D8_ASD_384','D8_ASD_494',
             'D8_Ctrl_290','D8_Ctrl_311','D8_Ctrl_317')
donor_sex <- factor(
    c('D8_Ctrl_290'='Male','D8_Ctrl_311'='Female','D8_Ctrl_317'='Female',
      'D8_ASD_375'='Male','D8_ASD_384'='Male','D8_ASD_494'='Male')[samples],
    levels=c('Female', 'Male'))

### **PALETTES**

# Cell types
cell.type.colors <- c(
 # '#1f77b4',
 '#ff7f0e',
 '#279e68',
 '#d62728',
 '#aa40fc',
 '#8c564b',
 '#e377c2',
 '#7f7f7f',
 '#b5bd61',
 '#17becf'
)
cell.type.colors <- c(
    '#F0E442',
    '#009E73',
    '#E69F00',
    '#D55E00',
    '#56B4E9',
    '#0072B2',
    '#CC79A7'
)


cell.type.colors <- setNames(cell.type.colors[1:length(cell.types)], cell.types)

# Cell cycle colors
phases.colors <- c(G1='gold', S='cadetblue', G2M='violet')

# Donors
samples.colors <- setNames(
    brewer.pal(n = length(samples), name = 'RdBu')[c(1:3,6:4)],
    samples
)

# Sex
origin_sex.colors <- c('Male'='cadetblue', 'Female'='gold')

scales::show_col(cell.type.colors)
cell.type.colors

# Export data

rm(mart)

save(list = ls(), file = '../data/commons.rda')

load('../data/commons.rda', verbose = T)

### Quick tests
- Genes ASD-associated with

library(dplyr)
(as.data.frame(imprinted.df)%>%subset(Imprinted.Status=='Imprinted'))$gene -> imp.genes.sure
(as.data.frame(imprinted.df)%>%subset(Imprinted.Status%in%c('Imprinted', 'Predicted')))$gene -> imp.genes.sure_or_pred

data.table::fread('../../cortical_disorders2/data/gene_disorder_associations.original.csv', data.table=F) %>% rename(GENE=V1) -> dis.df

asd.cols <- grep('ASD|SFAR', colnames(dis.df), value=T)

for (type in c('Imprinted', 'Imp+Predicted')){
    
    imp.genes <- if (type == 'Imprinted'){
        imp.genes.sure
    }else{
        imp.genes.sure_or_pred
    }

    print(glue::glue('\n==============================================='))
    print(glue::glue('\n==================\n{type}\n==================='))
    print(glue::glue('\n==============================================='))
    
    for (asd.dis in asd.cols){
        asd.genes <- dis.df$GENE[as.logical(dis.df[,asd.dis])]
        all.genes <- unique(c(asd.genes, imp.genes))
        TasdTimp <- sum((all.genes%in%asd.genes) & (all.genes%in%imp.genes))
        TasdFimp <- sum((all.genes%in%asd.genes) & (!all.genes%in%imp.genes))
        FasdTimp <- sum((!all.genes%in%asd.genes) & (all.genes%in%imp.genes))
        FasdFimp <- sum((!all.genes%in%asd.genes) & (!all.genes%in%imp.genes))

        print(glue::glue('\n=======================================\n{asd.dis}\n======================================='))
        print(glue::glue('ASD\tIMP\t'))
        print(glue::glue('T\tT\t{TasdTimp}'))
        print(glue::glue('T\tF\t{TasdFimp}'))
        print(glue::glue('F\tT\t{FasdTimp}'))
        print(glue::glue('F\tF\t{FasdFimp}'))
        print(glue::glue('\nImprinted and in ASD list:\n{paste(all.genes[(all.genes%in%asd.genes) & (all.genes%in%imp.genes)], collapse=" ")}'))


        cont.mat <- matrix(c(
            TasdTimp, TasdFimp,
            FasdTimp, FasdFimp), byrow=T, nrow=2)

        print(glue::glue('\nContingency table\n------------------'))
        print(cont.mat)
        print(glue::glue('------------------'))

        t <- fisher.test(cont.mat, alternative='greater')
        print(t)    
    }
}
